+++ server/controllers/api.js
@@ -7,61 +7,61 @@
  **********************************************************************************************************************/
 const fs = require("fs"); // https://nodejs.org/api/fs.html
 const path = require("path"); // https://nodejs.org/api/path.html
-const {promisify} = require("util");
-const readFileAsync = promisify(fs.readFile);
+const chalk = require("chalk"); // https://www.npmjs.com/package/chalk
 const moment = require("moment"); // https://momentjs.com/
 const mongoose = require("mongoose"); // http://mongoosejs.com/
 const Game = mongoose.model("Game");
 const logger = require("../handlers/logger/console");
 const cfg = require("../config");
 
-const areas = ["empire"];
-const locales = cfg.app.locales.map(locale => locale.name);
-// require all client language files. this is so node can cache the file contents
-const allTexts = [];
 
+/*
+ * prepare an array of text strings from filesystem
+ * this array is kept in memory and can be accessed fast by the api endpoint
+ */
+const allTexts = [];
+const locales = cfg.app.locales.map(locale => locale.name);
+// get filenames of existing area message files
+const areaLanguageFileNames = fs
+    .readdirSync(path.join(cfg.app.projectDir, "client", "lang", locales[0]))
+    .filter(fileName => fileName !== "common.json");
 // read client text files once and cache them
-areas.forEach(area => {
-    locales.forEach(locale => {
+locales.forEach(locale => {
+    const pathCommon = path.join(
+        cfg.app.projectDir,
+        "client",
+        "lang",
+        locale,
+        "common.json"
+    );
+    let commonTime = fs.statSync(pathCommon).mtime;
+    let commonContents = JSON.parse(fs.readFileSync(pathCommon, "utf-8"));
+    areaLanguageFileNames.forEach(areaFileName => {
         const pathArea = path.join(
             cfg.app.projectDir,
             "client",
             "lang",
             locale,
-            area + ".json"
+            areaFileName
         );
-        const pathCommon = path.join(
-            cfg.app.projectDir,
-            "client",
-            "lang",
-            locale,
-            "common.json"
-        );
-        let areaTime, commonTime;
-        let areaContents,
-            commonContents = {};
+        let areaTime;
+        let areaContents;
         try {
             areaContents = JSON.parse(fs.readFileSync(pathArea, "utf-8"));
-            areaTime = fs.statSync(pathArea).mtime;
-        } catch (e) {
-            logger.error(e);
-        }
-        try {
-            commonContents = JSON.parse(fs.readFileSync(pathCommon, "utf-8"));
             areaContents.common = commonContents;
-            commonTime = fs.statSync(pathCommon).mtime;
+            areaTime = fs.statSync(pathArea).mtime;
         } catch (e) {
             logger.error(e);
         }
         areaContents.version =
             moment(areaTime).diff(commonTime) > 0 ? areaTime : commonTime;
         allTexts.push({
-            slug: `txt-${locale}-${area}`,
+            slug: `txt-${locale}-${areaFileName.replace(".json","")}`,
             data: areaContents
         });
     });
 });
-logger.info(`[App] cached ${allTexts.length} client text objects.`);
+logger.info(`[App] cached ${chalk.yellow(allTexts.length)} client text objects.`);
 
 /*
  * get game status
@@ -114,6 +114,6 @@ exports.getTextStrings = async (req, res) => {
     if (textObj.data) {
         res.json(textObj.data);
     } else {
-        res.json({"error":"could not get client text strings."});
+        res.json({error: "could not get client text strings."});
     }
 };