+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+`26.12.2017`
+- [x] Changed api/textstrings so all texts are now cached on server start, api calls to the endpoint parse, filter and deliver the array simply.
+- [x] `Admin > Users` and `Admin > Games`: Added `moment.fromNow()` to datetimne displays.
+
 `23.12.2017`
 - [x] Boilerplating text string handling. language is passed via template from server to client, as well as the version of the json language files on disk. The client then checks if cached texts exists and if they are not older then the server files. If not, client asks server for texts and updates cache. Finally, vue ist extended and $txt is used for all text-strings for use in all components.
 You can now switch languages, and the header in "Empire" screen is dynamically updated.